Modularized: Borders class.
Update: All current classes within the program have successfully been refactored and modularized.
