+++ testBox2DSDL/Borders.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "Square.h"
+
+class Borders : public Square
+{
+    Square* bords;
+
+public:
+
+    void addBorders();
+
+    ~Borders();
+};
