+++ testBox2DSDL/Car.h
@@ -1,7 +1,4 @@
 #pragma once
-#include <SDL.h>
-#include <SDL_image.h>
-#include <box2d.h>
 #include "Square.h"
 #include "Circle.h"
