+++ testBox2DSDL/Circle.h
@@ -1,7 +1,4 @@
 #pragma once
-#include <SDL.h>
-#include <SDL_image.h>
-#include <box2d.h>
 #include "Shape.h"
 
 class Circle : public Shape
