+++ testBox2DSDL/Shape.h
@@ -1,4 +1,10 @@
 #pragma once
+#include <SDL.h>
+#include <SDL_image.h>
+#include <box2d.h>
+#include "InitialSetup.h"
+#include "Constants.h"
+
 class Shape
 {
 protected:
