+++ testBox2DSDL/Square.h
@@ -1,7 +1,4 @@
 #pragma once
-#include <SDL.h>
-#include <SDL_image.h>
-#include <box2d.h>
 #include "Shape.h"
 
 class Square : public Shape
